# workflow-service-sdk/api/openapi.yaml
openapi: 3.0.1
info:
  description: This is the API documentation for the Parodos Workflow Service. It
    provides operations to execute assessments to determine infrastructure options
    (tooling + environments). Also executes infrastructure task workflows to call
    downstream systems to stand-up an infrastructure option.
  title: Parodos Workflow Service API
  version: v1.0.0
servers:
- description: Generated server url
  url: http://localhost:8080
...
paths:
...
  /api/v1/projects:
    get:
      operationId: getProjects
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProjectResponseDTO'
                type: array
          description: Succeeded
        "304":
          description: Not Modified
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Not Found
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Internal Server Error
      summary: Returns a list of project
      tags:
      - Project
      x-accepts: application/json
    post:
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequestDTO'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponseDTO'
          description: Created
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Not Found
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Internal Server Error
      summary: Creates a new project
      tags:
      - Project
      x-content-type: application/json
      x-accepts: application/json
...
  /api/v1/projects/{id}:
    get:
      operationId: getProjectById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponseDTO'
          description: Succeeded
        "304":
          description: Not Modified
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          content: {}
          description: Not found
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Internal Server Error
      summary: Returns information about a specified project
      tags:
      - Project
      x-accepts: application/json
  /api/v1/projects/{id}/access:
    post:
      operationId: createAccessRequestToProject
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequestDTO'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessResponseDTO'
          description: Succeeded
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          content: {}
          description: Not found
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Internal Server Error
      summary: Request user access to project
      tags:
      - Project
      x-content-type: application/json
      x-accepts: application/json
  /api/v1/projects/{id}/users:
    delete:
      operationId: removeUsersFromProject
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              items:
                type: string
              type: array
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectUserRoleResponseDTO'
          description: Succeeded
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          content: {}
          description: Not found
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Internal Server Error
      summary: Remove users from project
      tags:
      - Project
      x-content-type: application/json
      x-accepts: application/json
    post:
      operationId: updateUserRolesToProject
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/UserRoleRequestDTO'
              type: array
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectUserRoleResponseDTO'
          description: Succeeded
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          content: {}
          description: Not found
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Conflict
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorMessageDTO'
          description: Internal Server Error
      summary: Update user roles in project
      tags:
      - Project
      x-content-type: application/json
      x-accepts: application/json
...
components:
  schemas:
    AccessRequestDTO:
      example:
        role: OWNER
        username: username
      properties:
        role:
          enum:
          - OWNER
          - DEVELOPER
          - ADMIN
          type: string
        username:
          type: string
      type: object
    AccessResponseDTO:
      example:
        approvalSentTo:
        - approvalSentTo
        - approvalSentTo
        accessRequestId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        escalationSentTo: escalationSentTo
        project:
          createdDate: 2000-01-23T04:56:07.000+00:00
          createdBy: createdBy
          name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        accessRequestId:
          format: uuid
          type: string
        approvalSentTo:
          items:
            type: string
          type: array
        escalationSentTo:
          type: string
        project:
          $ref: '#/components/schemas/ProjectDTO'
      type: object
...
    ErrorMessageDTO:
      properties:
        date:
          format: date-time
          type: string
        description:
          type: string
        message:
          type: string
        status:
          format: int32
          type: integer
      type: object
    ProjectDTO:
      example:
        createdDate: 2000-01-23T04:56:07.000+00:00
        createdBy: createdBy
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        createdBy:
          type: string
        createdDate:
          format: date-time
          type: string
        id:
          format: uuid
          type: string
        name:
          type: string
      type: object
    ProjectRequestDTO:
      example:
        name: name
        description: description
      properties:
        description:
          type: string
        name:
          type: string
      type: object
    ProjectResponseDTO:
      example:
        createdDate: 2000-01-23T04:56:07.000+00:00
        createdBy: createdBy
        accessRole: accessRole
        modifiedDate: 2000-01-23T04:56:07.000+00:00
        name: name
        description: description
        modifiedBy: modifiedBy
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        accessRole:
          type: string
        createdBy:
          type: string
        createdDate:
          format: date-time
          type: string
        description:
          type: string
        id:
          format: uuid
          type: string
        modifiedBy:
          type: string
        modifiedDate:
          format: date-time
          type: string
        name:
          type: string
      type: object
    ProjectUserRoleResponseDTO:
      example:
        userResponseDTOList:
        - roles:
          - OWNER
          - OWNER
          username: username
        - roles:
          - OWNER
          - OWNER
          username: username
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        projectName: projectName
      properties:
        id:
          format: uuid
          type: string
        projectName:
          type: string
        userResponseDTOList:
          items:
            $ref: '#/components/schemas/UserRoleResponseDTO'
          type: array
      type: object
    UserRoleRequestDTO:
      example:
        roles:
        - OWNER
        - OWNER
        username: username
      properties:
        roles:
          items:
            enum:
            - OWNER
            - DEVELOPER
            - ADMIN
            type: string
          type: array
        username:
          type: string
      type: object
    UserRoleResponseDTO:
      example:
        roles:
        - OWNER
        - OWNER
        username: username
      properties:
        roles:
          items:
            enum:
            - OWNER
            - DEVELOPER
            - ADMIN
            type: string
          type: array
          uniqueItems: true
        username:
          type: string
      type: object
...